Use Short instead of int to match Tasks table

+++ src/Databases/WorkflowDatabase.EF/Models/Task.cs
@@ -6,9 +6,9 @@ namespace WorkflowDatabase.EF.Models
     {
         public int TaskId { get; set; }
         public int WorkflowProcessId { get; set; }
-        public int DaysToDmEndDate { get; set; }
+        public short DaysToDmEndDate { get; set; }
         public DateTime DmEndDate { get; set; }
-        public int DaysOnHold { get; set; }
+        public short DaysOnHold { get; set; }
         public string RsdraNo { get; set; }
         public string SourceName { get; set; }
         public string Workspace { get; set; }
